+++ .jshintrc
@@ -1,0 +1,4 @@
+{
+    "asi": true,
+    "esversion": 8
+}
