+++ README.md
@@ -36,8 +36,6 @@ OR
 
 You can submit a related 'reply' response by selecting a node, then typing in the text area, then hitting submit.
 
-Once you add all your thoughts, do a git commit and push to github!
-
 The conversation graph is stored in `conversation-graph/conversation-graph.json`.
