+++ conversation-graph/index.html
@@ -62,7 +62,7 @@
     <div id="authors">Act as:</div>
     <div id="updates">New updates<button id="reload-button">Reload</button></div>
 
-    <script src="conversation-graph.js"></script>
+    <script src="convo-graph.bundle.js"></script>
 </body>
 
 </html>
