+++ editor-server.js
@@ -7,10 +7,6 @@ if (!shell.which('git')) {
     shell.exit(1);
 }
 
-// import our layout algorithm
-const { getLayoutForData } = require('./src/index')
-const cytoscapeConverter = require('./cytoscape-converter')
-
 const express = require('express')
 const app = express()
 const expressWs = require('express-ws')(app)
@@ -27,24 +23,6 @@ app.use(express.static('conversation-graph'))
 const ID = () => {
     return '_' + Math.random().toString(36).substr(2, 9)
 }
-
-// take a node list and edge list
-// and run it through our layout algorithm,
-// then convert that into a graph vis friendly format
-// and inject that into the JS file which is referenced in the
-// main index.html file
-const refreshJs = () => {
-    const { nodes, edges } = JSON.parse(fs.readFileSync('./conversation-graph/conversation-graph.json', 'utf-8'))
-    const focalTopicId = "_z0k58fv18"
-    const focalCoords = { x: 0, y: 0 }
-    const positions = getLayoutForData(nodes, edges, focalTopicId, focalCoords)
-    const cytoscapeData = cytoscapeConverter(nodes, edges, positions)
-    const js = fs.readFileSync('./conversation-graph/conversation-graph.template.js', 'utf-8')
-    let newjs = js.replace(/{{ data }}/gim, `${JSON.stringify(cytoscapeData)}`)
-    fs.writeFileSync('./conversation-graph/conversation-graph.js', newjs, 'utf-8')
-}
-// initialize the js file on startup, based on the graph we have
-refreshJs()
 
 /* POLLING of the remote git repository
     for updates. can be stopped, started,
@@ -78,10 +56,6 @@ const createPollTimer = (newInterval) => {
             // don't bother to send a message to the UI
             // if the change doesn't affect it
             if (matches.length === 0) return
-
-            // since there's updates
-            // refresh the js
-            refreshJs()
 
             // push notification to the client
             // letting it know the HTML has been updated
@@ -145,9 +119,6 @@ app.post('/add-node', (req, res) => {
     }
     fs.writeFileSync('./conversation-graph/conversation-graph.json', JSON.stringify({ nodes, edges }, null, 4), 'utf-8')
 
-    // update the js based on the new graph
-    refreshJs()
-
     // respond with the node id so that the UI
     // can trigger a page refresh, with the newly minted
     // node in focus
@@ -160,13 +131,6 @@ app.post('/add-node', (req, res) => {
         shell.exec(`git commit -m "node:${node.id}: ${node.text}"`, { silent: true })
         shell.exec('git push', { silent: true })
     }, 1000)
-})
-
-// a route to manually trigger a rebuild of the JS file
-// based on the current graph
-app.get('/refresh', (req, res) => {
-    refreshJs()
-    res.sendStatus(200)
 })
 
 app.listen(port, () => console.log(`Editor app listening on port ${port}!`))
