+++ package.json
@@ -23,5 +23,9 @@
     "mocha": "^6.0.2",
     "shelljs": "^0.8.3",
     "sinon": "^7.3.1"
+  },
+  "devDependencies": {
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.0"
   }
 }
